refactor: Archivos de productos renombrados
