+++ src/types/productTypes.ts
@@ -1,5 +1,5 @@
 import z from 'zod';
-import { productSchema } from '../schemas/product';
+import { productSchema } from '../schemas/productSchema';
 
 export type Product = z.infer<typeof productSchema>;
 export type ProductWithoutId = Omit<Product, 'id'>;
